Refine: Navigation bar refactor
